Split out test_matching.py
